Extract StakeStep component
